refactor: split violations_parsing by data source (jsonl, stream, evidence)
